refactor: Tag 리팩토링(#353)
