--- src/main/java/project/trendpick_pro/domain/tags/tag/entity/dto/response/TagResponse.java
@@ -1,13 +1,0 @@
-package project.trendpick_pro.domain.tags.tag.entity.dto.response;
-
-import lombok.Getter;
-
-@Getter
-public class TagResponse {
-
-    String tag;
-
-    public TagResponse(String tag) {
-        this.tag = tag;
-    }
-}
